// TUI_MONITORING/theme.d.ts
export declare const THEME: {
    bg: string;
    borderFg: string;
    info: string;
    sparkline: string;
    sparklineLbl: string;
};
export declare const ICONS: {
    bullet: string;
    running: string;
    queued: string;
    done: string;
    failed: string;
    error: string;
    warn: string;
    ok: string;
    up: string;
    down: string;
};